Rename ethernet/shared.rs -> ethernet/eth.rs
Since we plan rework the PHY code, calling this 'shared' doesn't make sense to
me
